refactor: unify theme service interface
