+++ app/lib/services/theme/theme_provider_factory.dart
@@ -4,13 +4,18 @@ import 'package:iwfpapp/services/theme/base_theme_provider.dart';
 import 'package:iwfpapp/services/theme/mock_theme_provider.dart';
 import 'package:iwfpapp/services/theme/offline_theme_provider.dart';
 
+/// App Theme Getter
+/// 
+/// This is a utility wrapper that abstract the creation of
+/// theme service provider. Code outside the theme directory
+/// should only use this method to access theme service.
 AppTheme getAppTheme(AppThemeType type,
     {ThemeType theme = ThemeType.WHITE, bool useSystem = true}) {
   switch (type) {
     case AppThemeType.OFFLINE:
       return OfflineAppTheme(themeType: theme, useSystem: useSystem);
     case AppThemeType.MOCK:
-      return MockThemeProvider(themeType: theme, useSystem: useSystem);
+      return MockAppTheme(themeType: theme, useSystem: useSystem);
     default:
       return OfflineAppTheme(themeType: theme, useSystem: useSystem);
   }
