+++ app/pubspec.yaml
@@ -11,7 +11,7 @@ description: A utility to help credit card users maximize cashback reward across
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.1.17+58
+version: 1.1.18+59
 
 environment:
   sdk: ">=2.1.0 <3.0.0"
