+++ app/test/integration/smoke_test.dart
@@ -21,7 +21,8 @@ void main() {
             create: (context) => getAppAuth(AppAuthType.MOCK_AUTH)),
         ChangeNotifierProvider(
             create: (context) => getDataBackend(BackendType.MOCK)),
-        ChangeNotifierProvider(create: (context) => AppContext(allowDynamicLink: false)),
+        ChangeNotifierProvider(
+            create: (context) => AppContext(allowDynamicLink: false)),
       ],
       child: IwfpApp(),
     ));
